Refactor UI styles and enhance service interfaces
Removed redundant CSS styles for a streamlined design, focusing on transitions and hover effects. Updated Razor components to include floating labels for improved user experience.

Significant modifications to C# service interfaces include clearer documentation and method restructuring. New interfaces added to support expanded functionality, emphasizing customizable user input.

Overall, these changes improve usability and maintainability across the application.
