+++ feedbackwebapp/Components/ContentFeed/Forms/HackerNewsFeedForm.razor.css
@@ -3,24 +3,5 @@
 }
 
 .form-floating {
-    transition: all 0.3s ease-in-out;
     margin-bottom: 1.5rem;
-}
-
-.form-control {
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    border-color: var(--border-color);
-}
-
-.form-control:focus {
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    border-color: var(--warning-color);
-    box-shadow: 0 0 0 0.25rem rgba(var(--warning-color-rgb), 0.25);
-}
-
-.text-muted {
-    color: var(--text-secondary) !important;
-    opacity: 0.9;
 }
