+++ feedbackwebapp/Components/Feedback/Forms/YouTubeInputForm.razor.css
@@ -9,24 +9,14 @@
 }
 
 .youtube-input .form-check-input {
-    cursor: pointer;
     width: 1.2em;
     height: 1.2em;
-    border-color: var(--border-color);
-    background-color: var(--bg-secondary);
-}
-
-.youtube-input .form-check-input:checked {
-    background-color: var(--primary-color);
-    border-color: var(--primary-color);
-    box-shadow: 0 0 0 0.15rem rgba(var(--primary-color-rgb), 0.25);
+    cursor: pointer;
 }
 
 .youtube-input .form-check-label {
     font-size: 1.1rem;
     cursor: pointer;
-    color: var(--text-primary);
-    transition: all 0.3s ease;
 }
 
 .youtube-input .form-check-label:hover {
@@ -34,29 +24,5 @@
 }
 
 .youtube-input .form-floating {
-    transition: all 0.3s ease-in-out;
     margin-bottom: 1.5rem;
-}
-
-.youtube-input .form-control {
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    border-color: var(--border-color);
-}
-
-.youtube-input .form-control:focus {
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    border-color: var(--primary-color);
-    box-shadow: 0 0 0 0.25rem rgba(var(--primary-color-rgb), 0.25);
-}
-
-.text-muted {
-    color: var(--text-secondary) !important;
-}
-
-.invalid-feedback {
-    font-size: 0.875rem;
-    margin-top: 0.25rem;
-    color: var(--danger-color);
 }
