+++ feedbackwebapp/Components/Feedback/Results/AnalysisResults.razor.css
@@ -9,99 +9,16 @@
     padding: 0.5rem;
 }
 
-.card {
-    border-radius: var(--border-radius);
-    transition: transform 0.3s, box-shadow 0.3s;
-    overflow: hidden;
-}
-
-.card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 0.75rem 2rem rgba(var(--primary-color-rgb), 0.15);
-}
-
-.card-header {
-    background: var(--primary-gradient) !important;
-    padding: 1.25rem;
-    border-bottom: none;
-    border-radius: var(--border-radius) var(--border-radius) 0 0 !important;
-}
-
-.card-header .card-title {
-    color: #ffffff;
-    margin: 0;
-    font-weight: 600;
-    font-size: 1.35rem;
-    letter-spacing: 0.5px;
-}
-
-.card-body {
-    padding: 1.5rem;
-}
-
-.btn-light {
-    transition: all 0.3s ease;
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    border-color: var(--border-color);
-    border-radius: var(--border-radius-sm);
-    padding: 0.6rem 1.2rem;
-}
-
-.btn-light:hover {
-    background-color: var(--hover-bg);
-    transform: translateY(-2px);
-    color: var(--text-primary);
-    box-shadow: 0 0.25rem 0.75rem rgba(var(--primary-color-rgb), 0.1);
-}
-
+/* Custom button styles not in app.css */
 .btn-outline-light {
-    transition: all 0.3s ease;
     color: #ffffff;
     border-color: rgba(255, 255, 255, 0.5);
-    border-radius: var(--border-radius-sm);
-    padding: 0.6rem 1.2rem;
     position: relative;
     overflow: hidden;
 }
 
 .btn-outline-light:hover:not(:disabled) {
     background-color: rgba(255, 255, 255, 0.15);
-    transform: translateY(-3px);
-    box-shadow: 0 0.25rem 0.75rem rgba(0, 0, 0, 0.1);
-}
-
-.action-btn {
-    position: relative;
-    overflow: hidden;
-}
-
-.action-btn::after {
-    content: '';
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    width: 120%;
-    height: 0;
-    padding-bottom: 120%;
-    border-radius: 50%;
-    background-color: rgba(255, 255, 255, 0.1);
-    transform: translate(-50%, -50%) scale(0);
-    opacity: 0;
-    transition: transform 0.5s, opacity 0.5s;
-}
-
-.action-btn:hover::after {
-    transform: translate(-50%, -50%) scale(1);
-    opacity: 1;
-}
-
-.action-btn i {
-    transition: all 0.3s ease;
-}
-
-.action-btn:hover i {
-    transform: scale(1.2) rotate(5deg);
 }
 
 .btn-outline-light:disabled {
@@ -113,6 +30,7 @@
     box-shadow: 0 0 0 0.25rem rgba(255, 255, 255, 0.2);
 }
 
+/* Markdown specific styles */
 .markdown-body h1,
 .markdown-body h2,
 .markdown-body h3,
@@ -225,6 +143,7 @@
     background-color: var(--hover-bg);
 }
 
+/* Speak button specific styles */
 .btn-speak-active {
     background-color: rgba(255, 255, 255, 0.25);
 }
