+++ feedbackwebapp/Components/Feedback/Results/BlueSkyFeedbackResults.razor.css
@@ -1,18 +1,3 @@
 .comments-list {
     margin-bottom: 1rem;
 }
-
-::deep .post-container {
-    animation: fadeIn 0.5s ease-in-out forwards;
-}
-
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-        transform: translateY(10px);
-    }
-    to {
-        opacity: 1;
-        transform: translateY(0);
-    }
-}
