+++ feedbackwebapp/Components/Feedback/Results/HackerNewsResults.razor.css
@@ -1,21 +1,3 @@
-.card {
-    background-color: var(--bg-secondary);
-    border-color: var(--border-color);
-}
-
-.card-header {
-    background: var(--primary-gradient) !important;
-}
-
-.card-header .card-title {
-    color: #ffffff;
-    margin: 0;
-}
-
-.card-title {
-    color: var(--text-primary);
-}
-
 .article-item {
     border-bottom: 1px solid var(--border-color);
     padding-bottom: 1rem;
@@ -29,7 +11,6 @@
 .article-link {
     color: var(--primary-color);
     text-decoration: none;
-    transition: all 0.2s ease;
 }
 
 .article-link:hover {
@@ -42,57 +23,18 @@
 }
 
 .article-text {
-    color: var(--text-primary);
     font-size: 0.9rem;
-}
-
-.article-stats .badge {
-    transition: transform 0.2s ease;
 }
 
 .article-stats .badge:hover {
     transform: translateY(-1px);
 }
 
-.btn-outline-secondary {
-    color: var(--text-secondary);
-    border-color: var(--border-color);
-}
-
-.btn-outline-secondary:hover {
-    background-color: var(--hover-bg);
-    color: var(--text-primary);
-    border-color: var(--border-color);
-}
-
-.btn-outline-light {
-    border-color: var(--border-color);
-    color: var(--text-secondary);
-}
-
-.btn-outline-light:hover {
-    background-color: var(--hover-bg);
-    color: var(--text-primary);
-    border-color: var(--border-color);
-}
-
-.comment-item {
-    background-color: var(--bg-secondary);
-    border: 1px solid var(--border-color);
-    transition: all 0.2s ease;
-}
-
 .comment-item:hover {
-    background-color: var(--hover-bg);
     transform: translateX(2px);
 }
 
-.comments-container {
-    animation: slideDown 0.3s ease-out;
-}
-
 .markdown-body {
-    color: var(--text-primary);
     background-color: transparent !important;
 }
 
@@ -102,17 +44,7 @@
 .markdown-body h4,
 .markdown-body h5,
 .markdown-body h6 {
-    color: var(--text-primary);
     border-bottom: 1px solid var(--border-color);
-}
-
-.markdown-body p {
-    color: var(--text-primary);
-}
-
-.markdown-body ul,
-.markdown-body ol {
-    color: var(--text-primary);
 }
 
 @keyframes slideDown {
