+++ feedbackwebapp/Components/Layout/MainLayout.razor.css
@@ -1,24 +1,11 @@
-/* Global theme utility styles - include this at the top of each component CSS file */
-
 .page {
     min-height: 100vh;
-    background-color: var(--bg-primary);
-    transition: background-color 0.3s ease;
 }
 
 .navbar {
-    background: var(--primary-gradient) !important;
-    box-shadow: 0 0.25rem 1rem rgba(var(--primary-color-rgb), 0.2);
-    transition: background-color 0.3s ease;
-    border-radius: 0 0 var(--border-radius) var(--border-radius);
     margin: 0 1rem 1.5rem;
     width: calc(100% - 2rem);
-}
-
-.navbar-brand {
-    font-weight: 600;
-    transition: transform 0.2s ease;
-    letter-spacing: 0.5px;
+    border-radius: 0 0 var(--border-radius) var(--border-radius);
 }
 
 .navbar-brand:hover {
@@ -31,10 +18,6 @@
 
 .nav-link {
     position: relative;
-    transition: all 0.3s ease;
-    padding: 0.5rem 1rem !important;
-    color: rgba(255, 255, 255, 0.85) !important;
-    border-radius: var(--border-radius-sm);
     margin: 0 0.25rem;
 }
 
@@ -42,19 +25,12 @@
     transition: transform 0.3s ease;
 }
 
-.nav-link:hover {
-    color: white !important;
-    background-color: rgba(255, 255, 255, 0.15);
-}
-
 .nav-link:hover i {
     transform: translateY(-2px) scale(1.1);
 }
 
 .nav-link.active {
-    color: white !important;
     font-weight: 500;
-    background-color: rgba(255, 255, 255, 0.2);
 }
 
 .nav-link.active::after {
@@ -82,12 +58,6 @@
 
 /* Responsive adjustments */
 @media (max-width: 768px) {
-    .container {
-        width: 100%;
-        padding-left: 1rem;
-        padding-right: 1rem;
-    }
-
     .navbar-nav {
         margin-left: 0;
         margin-top: 0.5rem;
@@ -104,17 +74,13 @@
 }
 
 #blazor-error-ui {
-    background: var(--bg-secondary);
-    color: var(--text-primary);
     bottom: 0;
-    box-shadow: 0 -1px 2px rgba(0, 0, 0, 0.2);
     display: none;
     left: 0;
     padding: 0.6rem 1.25rem 0.7rem 1.25rem;
     position: fixed;
     width: 100%;
     z-index: 1000;
-    transition: all 0.3s ease;
 }
 
 #blazor-error-ui .dismiss {
@@ -122,9 +88,7 @@
     position: absolute;
     right: 0.75rem;
     top: 0.5rem;
-    color: var(--text-primary);
     opacity: 0.7;
-    transition: opacity 0.2s ease;
 }
 
 #blazor-error-ui .dismiss:hover {
