+++ feedbackwebapp/Components/Pages/History.razor.css
@@ -1,4 +1,3 @@
-
 /* Searchbox (input) dark mode fix */
 [data-theme='dark'] .filter-section .form-control,
 [data-theme='dark'] .filter-section .form-select,
@@ -15,7 +14,7 @@
 
 /* History.razor.css - Slimmed down, Bootstrap-aligned, and theme-respecting */
 
-/* Container & Title */
+/* History title styling */
 .history-title {
     font-size: 2.25rem;
     font-weight: 600;
@@ -40,12 +39,6 @@
 .loading-state, .empty-state {
     text-align: center;
     padding: 3rem 1.5rem;
-    color: var(--text-secondary);
-    background: var(--bg-secondary);
-    border-radius: var(--border-radius);
-    border: 1px solid var(--border-color);
-    box-shadow: 0 0.5rem 1.5rem rgba(var(--primary-color-rgb), 0.08);
-    transition: all 0.3s ease;
 }
 
 .empty-state i {
@@ -53,54 +46,11 @@
     opacity: 0.8;
     color: var(--primary-color);
     margin-bottom: 1rem;
-    animation: pulse 3s infinite ease-in-out;
 }
 
-@keyframes pulse {
-    0% { transform: scale(1); opacity: 0.8; }
-    50% { transform: scale(1.1); opacity: 1; }
-    100% { transform: scale(1); opacity: 0.8; }
-}
-
-/* Filter section tweaks */
+/* Filter section */
 .filter-section.card {
     margin-bottom: 1.5rem;
-    border-radius: var(--border-radius);
-    background-color: var(--bg-secondary);
-    border: 1px solid var(--border-color);
-    box-shadow: 0 0.25rem 1rem rgba(var(--primary-color-rgb), 0.1);
-    transition: all 0.3s ease;
-}
-
-.filter-section.card:hover {
-    box-shadow: 0 0.5rem 1.5rem rgba(var(--primary-color-rgb), 0.15);
-}
-
-.filter-section .card-body {
-    padding: 1.25rem;
-}
-
-.filter-section .form-control,
-.filter-section .form-select {
-    border-radius: var(--border-radius-sm);
-    border: 1px solid var(--border-color);
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    transition: all 0.3s ease;
-    padding: 0.6rem 1rem;
-}
-
-.filter-section .form-control:focus,
-.filter-section .form-select:focus {
-    border-color: var(--primary-color);
-    box-shadow: 0 0 0 0.25rem rgba(var(--primary-color-rgb), 0.25);
-    transform: translateY(-2px);
-}
-
-.filter-section .form-label {
-    color: var(--text-primary);
-    font-weight: 500;
-    margin-bottom: 0.5rem;
 }
 
 /* Service icon and badge */
@@ -109,11 +59,6 @@
     color: var(--primary-color);
     margin-right: 0.75rem;
     opacity: 0.9;
-    transition: all 0.3s ease;
-}
-
-.history-item:hover .service-icon {
-    transform: scale(1.2) rotate(5deg);
 }
 
 .source-badge {
@@ -121,42 +66,23 @@
     color: #fff;
     font-weight: 600;
     font-size: 0.75rem;
-    border-radius: var(--border-radius-sm);
     padding: 0.4em 0.8em;
-    transition: all 0.3s ease;
-    box-shadow: 0 2px 5px rgba(var(--primary-color-rgb), 0.3);
 }
 
-.history-item:hover .source-badge {
-    transform: translateY(-2px);
-    box-shadow: 0 4px 8px rgba(var(--primary-color-rgb), 0.4);
-}
-
-/* History item list adjustments */
-.history-list {
-    border-radius: var(--border-radius);
-    overflow: hidden;
-}
-
+/* History item styles */
 .history-item {
-    transition: all 0.3s ease;
     border-left: 4px solid transparent;
-    background-color: var(--bg-secondary);
-    border-bottom: 1px solid var(--border-color);
     padding: 1.25rem;
 }
 
 .history-item:hover {
     border-left-color: var(--primary-color);
-    background-color: var(--hover-bg);
     transform: translateX(5px);
 }
 
 .history-item-title {
     font-weight: 600;
-    color: var(--text-primary);
     margin-bottom: 0.5rem;
-    transition: all 0.3s ease;
 }
 
 .history-item:hover .history-item-title {
@@ -164,141 +90,20 @@
 }
 
 .history-item-date {
-    color: var(--text-secondary);
     font-size: 0.85rem;
+    color: var(--text-secondary);
 }
 
 .markdown-content {
-    color: var(--text-primary);
     line-height: 1.6;
     font-size: 1rem;
     padding: 0.5rem 0;
-}
-
-/* Pagination */
-.pagination {
-    margin-top: 1.5rem;
-    justify-content: center;
-}
-
-.pagination .page-item .page-link {
-    color: var(--text-primary);
-    background-color: var(--bg-secondary);
-    border-color: var(--border-color);
-    transition: all 0.3s ease;
-    border-radius: var(--border-radius-sm);
-    margin: 0 0.2rem;
-}
-
-.pagination .page-item.active .page-link {
-    background-color: var(--primary-color);
-    border-color: var(--primary-color);
-    color: white;
-}
-
-.pagination .page-item .page-link:hover:not(.active) {
-    background-color: var(--hover-bg);
-    color: var(--primary-color);
-    transform: translateY(-2px);
-}
-
-.pagination .page-item.disabled .page-link {
-    color: var(--text-secondary);
-    background-color: var(--bg-secondary);
-    border-color: var(--border-color);
-}
-
-/* Action Buttons */
-.action-btn {
-    position: relative;
-    overflow: hidden;
-    transition: all 0.3s ease;
-    border-radius: var(--border-radius-sm);
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0.4rem 0.75rem;
-}
-
-.action-btn::after {
-    content: '';
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    width: 120%;
-    height: 0;
-    padding-bottom: 120%;
-    border-radius: 50%;
-    background-color: rgba(var(--primary-color-rgb), 0.08);
-    transform: translate(-50%, -50%) scale(0);
-    opacity: 0;
-    transition: transform 0.5s, opacity 0.5s;
-}
-
-.action-btn:hover {
-    transform: translateY(-3px);
-    box-shadow: 0 4px 15px rgba(var(--primary-color-rgb), 0.15) !important;
-}
-
-.action-btn:hover::after {
-    transform: translate(-50%, -50%) scale(1);
-    opacity: 1;
-}
-
-.action-btn:active {
-    transform: translateY(0);
-}
-
-.action-btn i {
-    transition: all 0.3s ease;
-}
-
-.action-btn:hover i {
-    transform: scale(1.2) rotate(5deg);
-}
-
-.btn-outline-primary {
-    color: var(--primary-color);
-    border-color: var(--primary-color);
-}
-
-.btn-outline-primary:hover {
-    background: var(--primary-gradient);
-    border-color: var(--primary-color);
-    color: white;
-}
-
-.btn-outline-secondary {
-    color: var(--text-secondary);
-    border-color: var(--border-color);
-}
-
-.btn-outline-secondary:hover {
-    background-color: var(--hover-bg);
-    color: var(--text-primary);
-    border-color: var(--border-color);
 }
 
 /* Clear History Button */
 .clear-all-btn {
     background: linear-gradient(135deg, #e98a97, #d96674);
     border: none;
-    position: relative;
-    overflow: hidden;
-    transition: all 0.3s ease;
-}
-
-.clear-all-btn:hover {
-    transform: translateY(-3px);
-    box-shadow: 0 4px 15px rgba(217, 102, 116, 0.3) !important;
-}
-
-.clear-all-btn i {
-    transition: all 0.3s ease;
-}
-
-.clear-all-btn:hover i {
-    transform: scale(1.2) rotate(5deg);
 }
 
 /* Responsive adjustments */
