+++ feedbackwebapp/Components/Pages/Home.razor.css
@@ -1,18 +1,3 @@
-.alert {
-    transition: all 0.3s ease;
-    box-shadow: 0 0.25rem 1rem rgba(var(--primary-color-rgb), 0.1);
-    background-color: var(--bg-secondary);
-    border: none;
-    border-radius: var(--border-radius);
-    padding: 1.25rem;
-    position: relative;
-    overflow: hidden;
-}
-
-.alert-info {
-    border-left: 4px solid var(--info-color);
-}
-
 .alert::before {
     content: '';
     position: absolute;
@@ -29,19 +14,16 @@
     z-index: 1;
 }
 
+.alert-info {
+    border-left: 4px solid var(--info-color);
+}
+
 .spinner-border {
     color: var(--primary-color);
 }
 
 .small {
     opacity: 0.9;
-    color: var(--text-secondary);
-    font-weight: 500;
-}
-
-.min-vh-100 {
-    background-color: var(--bg-primary);
-    transition: background-color 0.3s ease;
 }
 
 /* Source selector styling */
@@ -67,20 +49,6 @@
     height: 3px;
     background: var(--primary-gradient);
     border-radius: 3px;
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-    .container {
-        padding-left: 1rem;
-        padding-right: 1rem;
-    }
-
-    .card {
-        margin-left: -0.5rem;
-        margin-right: -0.5rem;
-        border-radius: var(--border-radius-sm);
-    }
 }
 
 /* Animation for status updates */
