--- feedbackwebapp/Services/Interfaces/IAutoDataSourceFeedbackService.cs
@@ -1,6 +1,0 @@
-namespace FeedbackWebApp.Services.Interfaces;
-
-public interface IAutoDataSourceFeedbackService : IFeedbackService
-{
-    // No need for GetFeedback declaration since it's inherited from IFeedbackService
-}
