--- feedbackwebapp/Services/Interfaces/IBlueSkyFeedbackService.cs
@@ -1,10 +1,0 @@
-using FeedbackWebApp.Services.Interfaces;
-
-namespace FeedbackWebApp.Services.Feedback;
-
-/// <summary>
-/// Interface for BlueSky feedback service
-/// </summary>
-public interface IBlueSkyFeedbackService : IFeedbackService
-{
-}
