--- feedbackwebapp/Services/Interfaces/IManualFeedbackService.cs
@@ -1,7 +1,0 @@
-namespace FeedbackWebApp.Services.Interfaces;
-
-public interface IManualFeedbackService : IFeedbackService
-{
-    string CustomPrompt { get; set; }
-    string Content { get; set; }
-}
